views structure
