+++ src/components/scss/board.scss
@@ -7,5 +7,6 @@
 @media screen and(max-width: 768px){
     .board{
         height: auto;
+        min-height: 100vh;
     }
 }
